chore(chromes): consolidate AppChrome and AdminChrome into layouts/chromes/
- sections/app-chrome/AppChrome.tsx  → layouts/chromes/AppChrome.tsx
- layouts/admin/ClientLayout.tsx     → layouts/chromes/AdminChrome.tsx
- layouts/admin/Layout.tsx           → layouts/chromes/AdminChromeLayout.tsx

No logic changes; import sites updated accordingly.
