+++ web/src/app/app/layout.tsx
@@ -6,7 +6,7 @@ import { ProjectsProvider } from "@/providers/ProjectsContext";
 import { VoiceModeProvider } from "@/providers/VoiceModeProvider";
 import AppSidebar from "@/sections/sidebar/AppSidebar";
 import { RootLayout } from "@opal/layouts";
-import AppChrome from "@/sections/app-chrome/AppChrome";
+import AppChrome from "@/layouts/chromes/AppChrome";
 
 export interface LayoutProps {
   children: React.ReactNode;
