+++ web/src/layouts/chromes/AdminChromeLayout.tsx
@@ -1,7 +1,7 @@
 import { redirect } from "next/navigation";
 import type { Route } from "next";
 import { requireAdminAuth } from "@/lib/auth/requireAuth";
-import ClientLayout from "@/layouts/admin/ClientLayout";
+import AdminChrome from "@/layouts/chromes/AdminChrome";
 import { AnnouncementBanner } from "@/components/header/AnnouncementBanner";
 
 export interface LayoutProps {
@@ -18,9 +18,9 @@ export default async function Layout({ children }: LayoutProps) {
   }
 
   return (
-    <ClientLayout>
+    <AdminChrome>
       <AnnouncementBanner />
       {children}
-    </ClientLayout>
+    </AdminChrome>
   );
 }
